ChatMain comments
Good OOP would say that we should set up a student to say their response - getResponse could perhaps be named getNextResponse, but otherwise this piece looks good. The one issue I have with it is that you're allowing your driver program to determine how much the student has to say, instead of asking the student "hasNext", for example
Since students have already been checked for uniqueness in the HashSet, you don't need to check again in your group's collection of students.
The structure used to set up groups is overcomplicated - using two collections requires more in-depth knowledge/parsing to understand than simply making a Group class that contains students. This is part of the cause of your bug in ordering.
Of all comments, the ones that are the worst are "TODO"s left in code and code that is commented out - the latter because you should be able to find this again in version control, and you'll never be able to answer the question of "when do we delete that code comment?"