Update GuiSection.java
Your comment about the serial ID is good as a learning comment, but should not be left in turned in code (and probably not as a JavaDoc comment in any case).
	JavaDoc looks like /** */, where standard multiline comments are /* */ - use JavaDoc for things that explain the usage of values/methods to outsiders (therefore a private variable is typically not the right place for them)

One highly recommended practice in programming is to declare variables as close to their use as possible. This means that if at all possible, you should avoid using "fields" or "instance variables" for elements that are local scope *in usage*. That means even if it's something that belongs to the class as a whole in terms of actual implementation, if I'm only using the variable locally and I don't have to otherwise declare it in the classes scope, I should still only declare it locally. Since everything you have except chatWindow is used only in the GuiSection constructor, this means only chatWindow would be required to have a field variable.

Fields are split (remind me to talk about this) on whether to use variable scope parameters when not require or not - this means protected/private may not be required when working in a class in Java - my preference and the requirement for assignments will be to include them. No points lost this time, but you will lose points on it in later assignments.
(Note that you also had theCaret and theBox defined twice, which is part of why pulling those out of a field and into local scope is good - it lets you avoid mistakes like those)